refactor: move menu to ui
