+++ src/lib/albums/AlbumMenu.svelte
@@ -8,9 +8,9 @@
         rightClickedAlbum,
         rightClickedTracks,
     } from "../../data/store";
-    import Menu from "../menu/Menu.svelte";
-    import MenuDivider from "../menu/MenuDivider.svelte";
-    import MenuOption from "../menu/MenuOption.svelte";
+    import Menu from "../ui/menu/Menu.svelte";
+    import MenuDivider from "../ui/menu/MenuDivider.svelte";
+    import MenuOption from "../ui/menu/MenuOption.svelte";
     import {
         type EnricherResult,
         enrichArtistCountry,
