+++ src/lib/queue/TrackMenu.svelte
@@ -9,9 +9,9 @@
         rightClickedTrack,
         rightClickedTracks,
     } from "../../data/store";
-    import Menu from "../menu/Menu.svelte";
-    import MenuDivider from "../menu/MenuDivider.svelte";
-    import MenuOption from "../menu/MenuOption.svelte";
+    import Menu from "../ui/menu/Menu.svelte";
+    import MenuDivider from "../ui/menu/MenuDivider.svelte";
+    import MenuOption from "../ui/menu/MenuOption.svelte";
     import {
         enrichArtistCountry,
         findCountryByArtist,
