+++ src/lib/ui/menu/MenuDivider.svelte
@@ -2,7 +2,7 @@
 
 <style>
     hr {
-		    border-top: 1px solid var(--menu-divider);
+        border-top: 1px solid var(--menu-divider);
         border-bottom: none;
         border-right: none;
         border-left: none;
