+++ src/lib/views/InternetArchiveView.svelte
@@ -23,8 +23,8 @@
         iaSelectedItem,
         webPlayerVolume,
     } from "../../data/store";
-    import Menu from "../menu/Menu.svelte";
-    import MenuOption from "../menu/MenuOption.svelte";
+    import Menu from "../ui/menu/Menu.svelte";
+    import MenuOption from "../ui/menu/MenuOption.svelte";
     import { open } from "@tauri-apps/plugin-shell";
 
     let audios;
